requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
APScheduler==3.10.4
tqdm==4.66.1
cachetools==5.3.2
aiohttp==3.9.1
psycopg2-binary==2.9.9
asyncpg==0.29.0 